Move to Elastic namespace
