--- lib/elastic-app-search.rb
@@ -1,4 +1,0 @@
-require 'elastic-app-search/client'
-
-module ElasticAppSearch
-end
