--- lib/elastic-app-search/version.rb
@@ -1,3 +1,0 @@
-module ElasticAppSearch
-  VERSION = '0.6.0'
-end
