+++ lib/elastic/app-search.rb
@@ -1,0 +1,6 @@
+require 'elastic/app-search/client'
+
+module Elastic
+  module AppSearch
+  end
+end
