+++ lib/elastic/app-search/version.rb
@@ -1,0 +1,5 @@
+module Elastic
+  module AppSearch
+    VERSION = '0.6.0'
+  end
+end
