+++ spec/spec_helper.rb
@@ -2,7 +2,7 @@ require 'bundler/setup'
 require 'rspec'
 require 'webmock/rspec'
 require 'awesome_print'
-require 'elastic-app-search'
+require 'elastic/app-search'
 require 'config_helper'
 
 WebMock.allow_net_connect!
